// src/models/lancamento.model.ts
export interface Lancamento {
  id?: string; // Optional because it's generated by Firestore or on add
  descricao: string;
  valor: number;
  data: any; // Firestore Timestamp or ISO string. For simplicity, 'any' for now.
  tipo: 'receita' | 'despesa';
  tipoPagamento: string; // e.g., 'Cartão de Crédito', 'Débito', 'PIX', 'Dinheiro'
  usuarioId: string; // To link to the user who created it
  nomeUsuario?: string; // Denormalized for easier display, if needed
  categoria?: string; // e.g., 'Alimentação', 'Transporte', 'Lazer', 'Salário'
}